feat: start the process of moving the backends into their own module
