+++ src/backends/datafusion_backend.py
@@ -1,13 +1,15 @@
 # SPDX-License-Identifier: Apache-2.0
 """Provides access to Datafusion."""
+from contextlib import contextmanager
 from pathlib import Path
+from typing import Iterator
 
 import pyarrow as pa
-from substrait.gen.proto import plan_pb2
-
-from gateway.backends.backend import Backend
 from gateway.converter.rename_functions import RenameFunctionsForDatafusion
 from gateway.converter.replace_local_files import ReplaceLocalFilesWithNamedTable
+from substrait.gen.proto import plan_pb2
+
+from backends.backend import Backend
 
 
 # pylint: disable=import-outside-toplevel
@@ -26,10 +28,9 @@ class DatafusionBackend(Backend):
 
         self._connection = datafusion.SessionContext()
 
-    def execute(self, plan: plan_pb2.Plan) -> pa.lib.Table:
-        """Execute the given Substrait plan against Datafusion."""
-        import datafusion.substrait
-
+    @contextmanager
+    def adjust_plan(self, plan: plan_pb2.Plan) -> plan_pb2.Plan:
+        """Modify the given Substrait plan for use with Datafusion."""
         file_groups = ReplaceLocalFilesWithNamedTable().visit_plan(plan)
         registered_tables = set()
         for files in file_groups:
@@ -41,23 +42,29 @@ class DatafusionBackend(Backend):
         RenameFunctionsForDatafusion().visit_plan(plan)
 
         try:
-            plan_data = plan.SerializeToString()
-            substrait_plan = datafusion.substrait.substrait.serde.deserialize_bytes(plan_data)
-            logical_plan = datafusion.substrait.substrait.consumer.from_substrait_plan(
-                self._connection, substrait_plan
-            )
-
-            # Create a DataFrame from a deserialized logical plan.
-            df_result = self._connection.create_dataframe_from_logical_plan(logical_plan)
-            for column_number, column_name in enumerate(df_result.schema().names):
-                df_result = df_result.with_column_renamed(
-                    column_name,
-                    plan.relations[0].root.names[column_number]
-                )
-            return df_result.to_arrow_table()
+            yield plan
         finally:
             for table_name in registered_tables:
                 self._connection.deregister_table(table_name)
+
+    def _execute_plan(self, plan: plan_pb2.Plan) -> Iterator[plan_pb2.Plan]:
+        """Execute the given Substrait plan against Datafusion."""
+        import datafusion.substrait
+
+        plan_data = plan.SerializeToString()
+        substrait_plan = datafusion.substrait.substrait.serde.deserialize_bytes(plan_data)
+        logical_plan = datafusion.substrait.substrait.consumer.from_substrait_plan(
+            self._connection, substrait_plan
+        )
+
+        # Create a DataFrame from a deserialized logical plan.
+        df_result = self._connection.create_dataframe_from_logical_plan(logical_plan)
+        for column_number, column_name in enumerate(df_result.schema().names):
+            df_result = df_result.with_column_renamed(
+                column_name,
+                plan.relations[0].root.names[column_number]
+            )
+        return df_result.to_arrow_table()
 
     def register_table(
             self, name: str, location: Path, file_format: str = 'parquet'
